Fixed various bugs
Resolved the problem of crashing when opening a dataset with more than 2 independent variables. It still doesn't open, but it won't crash.

Fixed the problem of being unable to open datasets if the filepath contains an underscore

Changed the data selection dialogue to use the array_id instead of name so that it is unique even if it's uglier.